+++ crates/cairo-profiler/src/trace_reader.rs
@@ -12,6 +12,7 @@ use trace_data::{
     CallTrace, CallTraceNode, ContractAddress, EntryPointSelector, ExecutionResources, L1Resources,
 };
 
+mod function_stack_trace;
 mod function_trace_builder;
 pub mod functions;
